Move Eventable to a `Moves` submodule Currently this relates only to move events, so can be scoped under moves for now. It will evolve further as other types of event get added.
